requirements.txt
fastapi
uvicorn
boto3
google-cloud-storage
python-dotenv
